[libromdata] SystemRegion: Handle 3-character country codes.
Some newer country codes use 3 characters. The return type is now
uint32_t in order to accomodate this. Note that the country code
is LSB-aligned, so existing 2-character codes will "just work".

MegaDriveRegions: Use uint32_t for the country code.

+++ src/libromdata/SystemRegion.hpp
@@ -38,10 +38,16 @@ class SystemRegion
 
 	public:
 		/**
-		 * Get the system country code. (two-character ISO-3166)
-		 * @return Two-character ISO-3166 country code as a uint16_t, or 0 on error.
+		 * Get the system country code. (ISO-3166)
+		 * This will always be an uppercase ASCII value.
+		 *
+		 * NOTE: Some newer country codes may use 3-character abbreviations.
+		 * The abbreviation will always be aligned towards the LSB, e.g.
+		 * 'US' will be 0x00005553.
+		 *
+		 * @return ISO-3166 country code as a uint32_t, or 0 on error.
 		 */
-		static uint16_t getCountryCode(void);
+		static uint32_t getCountryCode(void);
 };
 
 }
